+++ out/extension.js
@@ -21100,7 +21100,7 @@ var FeedbackViewProvider = class {
     this._bridgeBroadcastTimer = setInterval(() => {
       post();
       attempts += 1;
-      if (attempts >= 30 && this._bridgeBroadcastTimer) {
+      if (attempts >= 6 && this._bridgeBroadcastTimer) {
         clearInterval(this._bridgeBroadcastTimer);
         this._bridgeBroadcastTimer = void 0;
       }